refactor shared out of cmds
